+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [2026-05-16 22:37] home/godmode/lib/parsers.py
 - [2026-05-16 22:38] home/godmode/engine/racer.py
 - [2026-05-16 22:39] data/data/com.termux/files/home/godmode/engine/scorer.py
+- [2026-05-16 22:41] home/godmode/lib/filter.py